datetime -> datetime.time

+++ Helper.py
@@ -34,7 +34,7 @@ def activitiesToState1(nodes, list_activities):
 
 # devolve a quantidade de Minutos necessários para relizar o trajeto
 def Travel_Time( travel_mult, x1, y1, x2, y2, gmaps):
-    return Distance_Calculator( x1, y1, x2, y2) * travel_mult # values_dict['TRAVEL_TIME']
+    return int(Distance_Calculator( x1, y1, x2, y2) * travel_mult) # values_dict['TRAVEL_TIME']
 
 '''
 def Travel_Time( travel_mult, x1, y1, x2, y2, gmaps):
